Move histogram into histogram.go.
